+++ common/src/main/java/net/xstopho/resource_backpacks/mixin/common/InventoryMenuMixin.java
@@ -1,5 +1,6 @@
 package net.xstopho.resource_backpacks.mixin.common;
 
+import net.minecraft.world.SimpleContainer;
 import net.minecraft.world.entity.player.Inventory;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.inventory.InventoryMenu;
@@ -7,7 +8,7 @@ import net.minecraft.world.inventory.MenuType;
 import net.minecraft.world.inventory.RecipeBookMenu;
 import net.minecraft.world.item.crafting.CraftingInput;
 import net.minecraft.world.item.crafting.CraftingRecipe;
-import net.xstopho.resource_backpacks.client.slot.SurvivalBackpackSlot;
+import net.xstopho.resource_backpacks.client.slot.BackpackSlot;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
@@ -22,6 +23,6 @@ public abstract class InventoryMenuMixin extends RecipeBookMenu<CraftingInput, C
 
     @Inject(at = @At("TAIL"), method = "<init>")
     public void init(Inventory playerInventory, boolean active, final Player player, CallbackInfo info) {
-        this.addSlot(new SurvivalBackpackSlot(playerInventory, 26, 62));
+        this.addSlot(new BackpackSlot(new SimpleContainer(1), player, 26, 62));
     }
 }