+++ CHANGELOG.md
@@ -1,2 +1,3 @@
-#### v0.12.6-BETA
-- backpack in your hand gets locked when opened from inventory
+#### v0.13.0-BETA
+- adding missing keybind settings to neoforge
+- adding Backpack Slot for Creative Menu